requirements.txt
astroid==3.3.9
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
dill==0.3.9
graphviz==0.20.3
idna==3.10
isort==6.0.1
mccabe==0.7.0
networkx==3.4.2
platformdirs==4.3.6
pycparser==2.22
pydeps==3.0.1
PyGithub==2.6.1
PyJWT==2.10.1
pylint==3.3.5
PyNaCl==1.5.0
requests==2.32.3
soupsieve==2.6
stdlib-list==0.11.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
